Port the line classes
